Try3

+++ src/main/java/com/driver/services/impl/ParkingLotServiceImpl.java
@@ -30,47 +30,56 @@ public class ParkingLotServiceImpl implements ParkingLotService {
     @Override
     public Spot addSpot(int parkingLotId, Integer numberOfWheels, Integer pricePerHour) {
         ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
-        Spot spot = new Spot();
-        SpotType spotType = null;
-        if(numberOfWheels == 2) spotType = SpotType.TWO_WHEELER;
-        else if(numberOfWheels == 4) spotType = SpotType.FOUR_WHEELER;
-        else spotType = SpotType.OTHERS;
-        spot.setSpotType(spotType);
-        spot.setOccupied(false);
-        spot.setPricePerHour(pricePerHour);
-        spot.setParkingLot(parkingLot);
-        parkingLot.getSpotList().add(spot);
+        Spot spot = null;
+        if(parkingLot != null){
+            spot = new Spot();
+            SpotType spotType = null;
+            if(numberOfWheels == 2) spotType = SpotType.TWO_WHEELER;
+            else if(numberOfWheels == 4) spotType = SpotType.FOUR_WHEELER;
+            else spotType = SpotType.OTHERS;
+            spot.setSpotType(spotType);
+            spot.setOccupied(false);
+            spot.setPricePerHour(pricePerHour);
+            spot.setParkingLot(parkingLot);
+            parkingLot.getSpotList().add(spot);
 
-       // parkingLotRepository1.save(parkingLot); // saves parkinglot and spot
-
-        return spotRepository1.save(spot);
+            parkingLotRepository1.save(parkingLot); // saves parkinglot and spot
+        }
+        return spot;
     }
 
     @Override
     public void deleteSpot(int spotId) {
         Spot spot = spotRepository1.findById(spotId).get();
-        ParkingLot parkingLot = spot.getParkingLot();
-        parkingLot.getSpotList().remove(spot);
-        spotRepository1.deleteById(spotId);
+        if(spot != null){
+            ParkingLot parkingLot = spot.getParkingLot();
+            parkingLot.getSpotList().remove(spot);
+            spotRepository1.deleteById(spotId);
+        }
     }
 
     @Override
     public Spot updateSpot(int parkingLotId, int spotId, int pricePerHour) {
         Spot spot = spotRepository1.findById(spotId).get();
-        ParkingLot parkingLot  = spot.getParkingLot();
-        parkingLot.getSpotList().remove(spot);
-        spot.setPricePerHour(pricePerHour);
-        parkingLot.getSpotList().add(spot);
-        return spotRepository1.save(spot);
+        if(spot != null){
+            ParkingLot parkingLot  = spot.getParkingLot();
+            parkingLot.getSpotList().remove(spot);
+            spot.setPricePerHour(pricePerHour);
+            parkingLot.getSpotList().add(spot);
+            return spotRepository1.save(spot);
+        }
+        return null;
     }
 
     @Override
     public void deleteParkingLot(int parkingLotId) {
         ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
-        List<Spot> spots = parkingLot.getSpotList();
-        for(Spot spot : spots){
-            spotRepository1.deleteById(spot.getId());
+        if(parkingLot != null){
+            List<Spot> spots = parkingLot.getSpotList();
+            for(Spot spot : spots){
+                spotRepository1.deleteById(spot.getId());
+            }
+            parkingLotRepository1.deleteById(parkingLotId);
         }
-        parkingLotRepository1.deleteById(parkingLotId);
     }
 }
